pyproject.toml
[tool.poetry]
name = "emailsender"
version = "0.1.0"
description = ""
authors = ["Oleksandr Nazarevych <oleksandr.o.nazarevych@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pydantic = "^2.10.5"
sqlalchemy = "^2.0.36"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.10"
markdown = "^3.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
